pyproject.toml
[tool.poetry]
name = "birgo"
version = "0.1.0"
description = ""
authors = ["Sirojiddin Mustafayev <sirojiddinmustafayev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.7"
python-dotenv = "^1.1.1"
dj-database-url = "^3.0.1"
psycopg2-binary = "^2.9.10"
pytelegrambotapi = "^4.29.1"
pylint-django = "^2.6.1"
init-data-py = "^0.2.6"
pillow = "^11.3.0"
humanize = "^4.13.0"

[tool.pylint]
ignore-paths = [
    "manage.py",  # Ignore all files within any 'migrations' directory
    ".*/migrations/.*",  # Ignore all files within any 'migrations' directory
    ".*\\.pyi$",         # Ignore all stub files (.pyi)
]
disable = [
    "C0114", # missing-module-docstring
    "W0641", # possibly-unused-variable
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
